# spec/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Api Documentation",
    "version" : "1.0",
    "title" : "Api Documentation",
    "termsOfService" : "urn:tos",
    "contact" : { },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host" : "localhost",
  "basePath" : "/",
...
  "paths" : {
...
    "/api/standard-price/{salesorg}/{salesoffice}" : {
      "get" : {
        "tags" : [ "standard-price-controller" ],
        "summary" : "getStdPricesForSalesOfficeAndSalesOrg",
        "operationId" : "getStdPricesForSalesOfficeAndSalesOrgUsingGET",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "salesoffice",
          "in" : "path",
          "description" : "salesoffice",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "salesorg",
          "in" : "path",
          "description" : "salesorg",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "zone",
          "in" : "query",
          "description" : "zone",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/MaterialStdPriceDTO"
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/standard-price/{salesorg}/{salesoffice}/{material}" : {
      "get" : {
        "tags" : [ "standard-price-controller" ],
        "summary" : "getStdPriceForSalesOrgAndSalesOfficeAndMaterial",
        "operationId" : "getStdPriceForSalesOrgAndSalesOfficeAndMaterialUsingGET",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "material",
          "in" : "path",
          "description" : "material",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "salesoffice",
          "in" : "path",
          "description" : "salesoffice",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "salesorg",
          "in" : "path",
          "description" : "salesorg",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "zone",
          "in" : "query",
          "description" : "zone",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/MaterialStdPriceDTO"
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
...
    "/api/v1/standard-price/{salesorg}/{salesoffice}" : {
      "get" : {
        "tags" : [ "standard-price-controller" ],
        "summary" : "getStdPricesForSalesOfficeAndSalesOrg",
        "operationId" : "getStdPricesForSalesOfficeAndSalesOrgUsingGET_1",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "salesoffice",
          "in" : "path",
          "description" : "salesoffice",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "salesorg",
          "in" : "path",
          "description" : "salesorg",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "zone",
          "in" : "query",
          "description" : "zone",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/MaterialStdPriceDTO"
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/v1/standard-price/{salesorg}/{salesoffice}/{material}" : {
      "get" : {
        "tags" : [ "standard-price-controller" ],
        "summary" : "getStdPriceForSalesOrgAndSalesOfficeAndMaterial",
        "operationId" : "getStdPriceForSalesOrgAndSalesOfficeAndMaterialUsingGET_1",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "material",
          "in" : "path",
          "description" : "material",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "salesoffice",
          "in" : "path",
          "description" : "salesoffice",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "salesorg",
          "in" : "path",
          "description" : "salesorg",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "zone",
          "in" : "query",
          "description" : "zone",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/MaterialStdPriceDTO"
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "MaterialDTO" : {
      "type" : "object",
      "properties" : {
        "accountingRequirement" : {
          "type" : "string"
        },
        "categoryDescription" : {
          "type" : "string"
        },
        "categoryId" : {
          "type" : "string"
        },
        "classDescription" : {
          "type" : "string"
        },
        "classId" : {
          "type" : "string"
        },
        "depth" : {
          "type" : "number",
          "format" : "double"
        },
        "dimensionUnit" : {
          "type" : "string"
        },
        "dimensions" : {
          "type" : "string"
        },
        "distributionChannel" : {
          "type" : "string"
        },
        "height" : {
          "type" : "number",
          "format" : "double"
        },
        "lastChangedDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "lastChangedTime" : {
          "type" : "string"
        },
        "lastChangedTimestamp" : {
          "type" : "string"
        },
        "manufacturerMaterialNumber" : {
          "type" : "string"
        },
        "material" : {
          "type" : "string"
        },
        "materialDescription" : {
          "type" : "string"
        },
        "materialExpired" : {
          "type" : "boolean"
        },
        "materialGroup" : {
          "type" : "string"
        },
        "materialGroupDescription" : {
          "type" : "string"
        },
        "materialType" : {
          "type" : "string"
        },
        "materialTypeDescription" : {
          "type" : "string"
        },
        "maxLoad" : {
          "type" : "number",
          "format" : "double"
        },
        "maxLoadUnit" : {
          "type" : "string"
        },
        "netWeight" : {
          "type" : "number",
          "format" : "double"
        },
        "salesOrganization" : {
          "type" : "string"
        },
        "subCategoryDescription" : {
          "type" : "string"
        },
        "subCategoryId" : {
          "type" : "string"
        },
        "volume" : {
          "type" : "number",
          "format" : "double"
        },
        "volumeUnit" : {
          "type" : "string"
        },
        "weightUnit" : {
          "type" : "string"
        },
        "width" : {
          "type" : "number",
          "format" : "double"
        }
      },
      "title" : "MaterialDTO"
    },
...
    "MaterialStdPriceDTO" : {
      "type" : "object",
      "properties" : {
        "currency" : {
          "type" : "string"
        },
        "designation" : {
          "type" : "string"
        },
        "deviceType" : {
          "type" : "string"
        },
        "formattedScaleQuantum" : {
          "type" : "string"
        },
        "formattedStandardPrice" : {
          "type" : "string"
        },
        "material" : {
          "type" : "string"
        },
        "materialData" : {
          "$ref" : "#/definitions/MaterialDTO"
        },
        "materialExpired" : {
          "type" : "string"
        },
        "materialType" : {
          "type" : "string"
        },
        "materialTypeDesignation" : {
          "type" : "string"
        },
        "pricingUnit" : {
          "type" : "string"
        },
        "productGroup" : {
          "type" : "string"
        },
        "productGroupDesignation" : {
          "type" : "string"
        },
        "quantumUnit" : {
          "type" : "string"
        },
        "salesOffice" : {
          "type" : "string"
        },
        "salesOrg" : {
          "type" : "string"
        },
        "scaleQuantum" : {
          "type" : "number",
          "format" : "double"
        },
        "standardPrice" : {
          "type" : "number",
          "format" : "double"
        },
        "validFrom" : {
          "type" : "string",
          "format" : "date-time"
        },
        "validTo" : {
          "type" : "string",
          "format" : "date-time"
        },
        "zone" : {
          "type" : "string"
        }
      },
      "title" : "MaterialStdPriceDTO"
    },
...
  }
}